+++ src/arch/host/include/arch/audio/eq_fir/eq_fir.h
@@ -1,0 +1,19 @@
+/* SPDX-License-Identifier: BSD-3-Clause
+ *
+ * Copyright(c) 2019 Intel Corporation. All rights reserved.
+ *
+ * Author: Tomasz Lauda <tomasz.lauda@linux.intel.com>
+ */
+
+#ifdef __SOF_AUDIO_EQ_FIR_H__
+
+#ifndef __ARCH_AUDIO_EQ_FIR_EQ_FIR_H__
+#define __ARCH_AUDIO_EQ_FIR_EQ_FIR_H__
+
+#endif /* __ARCH_AUDIO_EQ_FIR_EQ_FIR_H__ */
+
+#else
+
+#error "This file shouldn't be included from outside of sof/audio/eq_fir.h"
+
+#endif /* __SOF_AUDIO_EQ_FIR_H__ */
