+++ src/arch/xtensa/audio/eq_fir/CMakeLists.txt
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: BSD-3-Clause
+
+add_local_sources(sof fir_hifi2ep.c fir_hifi3.c)
