+++ src/audio/eq_fir/CMakeLists.txt
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
-add_local_sources(sof eq_fir.c fir_hifi2ep.c fir_hifi3.c fir.c)
+add_local_sources(sof eq_fir.c fir.c)
